enhancement: Moved sse_handler to utils, refactored is_first_flag_submission function so it doesnt count all flag_submissions and just fetchs any flag_submission for given task_id, moved checking for flag submissions for given task_id before inserting it
