+++ backend/src/routes/flag/submit.rs
@@ -2,9 +2,9 @@ use crate::entities::sea_orm_active_enums::TeamStatus;
 use crate::entities::{flag_capture, teams, users};
 use crate::routes::flag::AuthMiddleware;
 use crate::routes::flag::FlagError;
-use crate::routes::leaderboard::sse_handler::SseMessage;
 use crate::utils::app_state::AppState;
 use crate::utils::error::Error;
+use crate::utils::sse_handler::SseMessage;
 use actix_web::web::{Data, Json};
 use actix_web::{HttpResponse, post};
 use actix_web_validation::Validated;
