+++ backend/src/routes/leaderboard/mod.rs
@@ -1,15 +1,15 @@
+use crate::utils::sse_handler;
 use actix_web::http::StatusCode;
 use actix_web::{HttpResponse, ResponseError};
 use thiserror::Error;
 
 mod chart;
-pub mod sse_handler;
 mod teams;
 
 pub fn config(config: &mut utoipa_actix_web::service_config::ServiceConfig) {
     config.service(chart::chart);
     config.service(teams::teams);
-    cfg.route(
+    config.route(
         "/updates",
         actix_web::web::get().to(sse_handler::sse_handler),
     );
