+++ backend/src/utils/sse_handler.rs
